bot/requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
python-magic==0.4.27
SpeechRecognition==3.10.0
pydub==0.25.1
PyPDF2==3.0.1
python-docx==1.1.0
redis==5.0.1
aiosqlite==0.19.0
aiohttp==3.9.1
